move reporter and login to pages folder
