+++ frontend/src/App.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import { BrowserRouter as Router, Switch, Route, Redirect } from "react-router-dom";
-import { Login } from "./login";
-import { Reporter } from "./reporter";
+import { Login } from "./pages/login";
+import { Reporter } from "./pages/reporter";
 
 export const App = () => {
   return(
